backend/requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-heroku
djangorestframework==3.15.2
firebase-admin==6.5.0
gcloud>=0.18.3
gcloud>=0.18.3
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jws==0.1.3
msgpack==1.1.0
oauth2client>=4.1.2
oauth2client>=4.1.2
proto-plus==1.24.0
protobuf==5.29.0rc1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome>=3.6.4
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
python-jwt==2.0.1
requests>=2.18.0, <3.0.0
requests-toolbelt>=0.7.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3<2